pyproject.toml
[build-system]
requires = [
    "hatchling>=1.4.1",
]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "*.py"
]
exclude = [
    "device/",
    "sample/",
]

[tool.hatch.version]
path = "testbed/swarmit/__init__.py"

[project]
name = "swarmit"
dynamic = ["version"]
authors = [
    { name="Alexandre Abadie", email="alexandre.abadie@inria.fr" },
]
dependencies = [
    "click          == 8.1.7",
    "cryptography   == 43.0.1",
    "pydotbot       == 0.22.0",
    "pyserial       == 3.5",
    "rich           == 13.8.1",
    "structlog      == 24.4.0",
    "tqdm           == 4.66.5",
    "paho-mqtt      >= 2.1.0",
]
description = "Run Your Own Robot Swarm Testbed."
readme = "README.md"
license = { text="BSD" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
]

[project.urls]
"Homepage" = "https://github.com/DotBots/swarmit"
"Bug Tracker" = "https://github.com/DotBots/swarmit/issues"

[project.scripts]
swarmit = "testbed.cli.main:main"

[tool.ruff]
lint.select = ["E", "F"]
line-length = 88
lint.ignore = ["E501"]

[tool.isort]
multi_line_output = 3  # Use Vertical Hanging Indent
profile = "black"

[tool.black]
line-length = 79
skip-string-normalization = true
